+++ NEWS.md
@@ -1,6 +1,8 @@
 # mlr3data 0.2.1
 
 * Removed dependency on orphaned package `bibtex`.
+* Added optdigits data obtained via the `OpenML` package.
+* Added moneyball data obtained via the `OpenML` package.
 
 # mlr3data 0.2.0